Move ParseErrorCause to scylla-cql-core
It appears in CustomTypeParseError, which is a public error type.
