+++ scylla-cql-core/src/utils/mod.rs
@@ -1,1 +1,3 @@
 pub(crate) mod safe_format;
+
+pub mod parse;
