+++ scylla-cql/src/utils/parse.rs
@@ -1,7 +1,7 @@
 //! Simple general-purpose recursive-descent parser.
 //! Used for parsing strings in the CQL protocol.
 
-use std::fmt::Display;
+pub use scylla_cql_core::utils::parse::ParseErrorCause;
 
 /// An error that can occur during parsing.
 #[derive(Copy, Clone)]
@@ -22,26 +22,6 @@ impl ParseError {
     /// Returns the error cause.
     pub fn get_cause(&self) -> ParseErrorCause {
         self.cause
-    }
-}
-
-/// Cause of the parsing error.
-/// Should be lightweight so that it can be quickly discarded.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum ParseErrorCause {
-    /// Expected a specific string, but it was not found.
-    Expected(&'static str),
-    /// Other error, described by a string.
-    Other(&'static str),
-}
-
-impl Display for ParseErrorCause {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ParseErrorCause::Expected(e) => write!(f, "expected {e:?}"),
-            ParseErrorCause::Other(e) => f.write_str(e),
-        }
     }
 }
